upadate

+++ service.go
@@ -54,8 +54,10 @@ func (sn *storageNode) Service() {
 		log.Printf("[gconfig]配置更新重启矿机 %v\n", gc)
 		config.Gconfig.Save()
 		// 随机等待重启，错开高峰
-		os.Remove(path.Join(util.GetYTFSPath(), ".utp_params.json"))
-		os.Remove(path.Join(util.GetYTFSPath(), ".dtp_params.json"))
+		if config.Gconfig.Clean {
+			os.Remove(path.Join(util.GetYTFSPath(), ".utp_params.json"))
+			os.Remove(path.Join(util.GetYTFSPath(), ".dtp_params.json"))
+		}
 		time.Sleep(time.Duration(rand.Int63n(10)) * time.Second)
 		os.Exit(0)
 	}
